Specify `sublime.LITERAL` in `view.find`
By default `view.find` assumes the search expression is a regex, which is
not what we want since we are searching for substrings present in the
text. If the text has characters such as `(`, it wasn't working as
expected. Specify `sublime.LITERAL` flag to the `view.find` calls to
explicitly specify that it is not a regex.